chore: removed use of Java Optional in MailTemplateHelper.kt; extracted extension functions; extended unit tests
